whitespace changes and minor tweaks to get a minimal diff to original bicirik-dwarf source
