--- src/main/java/com/bicirikdwarf/elf/ElfSymbol.java
@@ -1,22 +1,0 @@
-package com.bicirikdwarf.elf;
-
-import java.nio.ByteBuffer;
-
-public class ElfSymbol {
-    public int st_name;
-    public int st_value;
-    public int st_size;
-    public byte st_info;
-    public byte st_other;
-    public short st_shndx;
-    public String symbol_name;
-
-    public void parse(ByteBuffer buffer) {
-        st_name = buffer.getInt();
-        st_value = buffer.getInt();
-        st_size = buffer.getInt();
-        st_info = buffer.get();
-        st_other = buffer.get();
-        st_shndx = buffer.getShort();
-    }
-}
